update: seperate unrelated code
