+++ pocsuite/api/rcGen.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+"""
+Copyright (c) 2014-2015 pocsuite developers (http://seebug.org)
+See the file 'docs/COPYING' for copying permission
+"""
+import os
+
+def initial():
+    currentUserHomePath = os.path.expanduser('~')
+    _ = """[zoomeye]\nusername = Your ZoomEye Username\npassword = Your ZoomEye Password\n\n[token]\nseebug = Your Seebug Token"""
+    if not os.path.isfile(currentUserHomePath + '/.pocsuiterc'):
+        with open(currentUserHomePath + '/.pocsuiterc', 'w') as fp:
+            fp.write(_)
